requirements.txt
streamlit
google-generativeai
git init